requirements.txt
torch>=2.2.0,<3.0.0
numpy>=1.25.0,<2.0.0
pandas>=2.0.0,<3.0.0
matplotlib>=3.8.0,<4.0.0
gymnasium>=0.29.0,<1.0.0
stable-baselines3>=2.0.0,<3.0.0
scikit-learn>=1.4.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
tqdm>=4.0.0,<5.0.0
tensorboard>=2.15.0,<3.0.0
seaborn>=0.13.0,<1.0.0
jupyter>=1.0.0,<2.0.0
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0